feat(kernel): add camera capture wire types for panel-RPC
Agent-Id: agent-a542fb16-473e-4f46-a57d-4f655e811c27
Linked-Note-Id: 22b2ddcb-89f3-47f9-a627-3fb756aaed3c
