+++ packages/webapp/src/kernel/panel-rpc.ts
@@ -43,6 +43,7 @@ import type { LeaderTrayRuntimeStatus } from '../scoops/tray-leader.js';
 import type { TrayLeaveResult } from '../scoops/tray-leave.js';
 import type { SudoDecision, SudoRequest } from '../sudo/types.js';
 import type { HidDeviceFilter, HidDeviceInfo } from './hid-device-registry.js';
+import type { CameraCaptureRequest, CameraCaptureResult } from './panel-rpc-camera-types.js';
 import type {
   SerialDeviceInfo,
   SerialFilter,
@@ -141,28 +142,7 @@ export type PanelRpcRequest =
     }
   | {
       op: 'capture-camera';
-      payload: {
-        mode: 'photo' | 'video';
-        deviceId?: string;
-        audioDeviceId?: string;
-        captureAudio?: boolean;
-        /**
-         * Open a video track on the stream. Defaults to true for
-         * photo / video mode; set to false for audio-only video
-         * captures so `getUserMedia` doesn't request a camera.
-         */
-        captureVideo?: boolean;
-        width?: number;
-        height?: number;
-        frameRate?: number;
-        exactSize?: boolean;
-        mimeType: string;
-        quality?: number;
-        durationMs?: number;
-        /** Photo mode: ms to let the sensor's auto-exposure settle
-         * before grabbing the frame. */
-        warmupMs?: number;
-      };
+      payload: CameraCaptureRequest;
     }
   | { op: 'enumerate-media-devices'; payload?: undefined }
   // Speech capture / transcription for the `hear` command. The kernel
@@ -551,13 +531,7 @@ export interface PanelRpcResults {
   'window-open': { opened: boolean };
   'oauth-popup': { redirectUrl: string | null };
   'silent-renew': { accessToken: string | null };
-  'capture-camera': {
-    bytes: ArrayBuffer;
-    mimeType: string;
-    width: number;
-    height: number;
-    durationMs?: number;
-  };
+  'capture-camera': CameraCaptureResult;
   'enumerate-media-devices': {
     videoinputs: Array<{ deviceId: string; label: string; groupId?: string }>;
     audioinputs: Array<{ deviceId: string; label: string; groupId?: string }>;
